move screening query into provider
